feat: feedback 로직 수정

+++ satellite-application/src/main/kotlin/kr/hs/entrydsm/satellite/domain/feedback/usecase/DeleteFeedbackUseCase.kt
@@ -3,6 +3,7 @@ package kr.hs.entrydsm.satellite.domain.feedback.usecase
 import kr.hs.entrydsm.satellite.common.annotation.UseCase
 import kr.hs.entrydsm.satellite.domain.document.exception.DocumentNotFoundException
 import kr.hs.entrydsm.satellite.domain.document.spi.DocumentPort
+import kr.hs.entrydsm.satellite.domain.feedback.exception.FeedbackNotFoundException
 import kr.hs.entrydsm.satellite.domain.feedback.spi.FeedbackPort
 import java.util.*
 
@@ -15,6 +16,9 @@ class DeleteFeedbackUseCase(
 
         documentPort.queryById(documentId) ?: throw DocumentNotFoundException
 
+        if (!feedbackPort.existsByDocumentIdAndElementId(documentId, elementId)) {
+            throw FeedbackNotFoundException
+        }
         feedbackPort.deleteByDocumentIdAndElementId(documentId, elementId)
     }
 }
